requirements.txt
aiofiles
asyncio
GitPython
openai
pinecone-client
fastapi
uvicorn
PyYAML
pytest
pytest-asyncio
